% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_regions}
\alias{check_regions}
\title{Check that ordered regions match original regions}
\usage{
check_regions(regions, ordered_regions)
}
\arguments{
\item{regions}{Vector of regions}

\item{ordered_regions}{Vector of ordered regions}
}
\value{
Printed informational statements
}
\description{
Check that ordered regions match original regions
}
