% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compile_taxonomy}
\alias{compile_taxonomy}
\title{Crunch the taxonomy}
\usage{
compile_taxonomy(files, regions, tax = rebird:::tax)
}
\arguments{
\item{files}{Vector of html file paths to crunch}

\item{regions}{Vector of region names to crunch, matching order of files}

\item{tax}{eBird taxonomy to use. Defaults to current taxonomy object from package
rebird}
}
\value{
Vector of species names found across the provided filters
}
\description{
Crunch the taxonomy
}
