% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crunch_filters}
\alias{crunch_filters}
\title{Create nested data structure and crunch the filters}
\usage{
crunch_filters(
  files,
  species,
  ordered_regions,
  filter_prefix,
  filter_suffix,
  tax = rebird:::tax
)
}
\arguments{
\item{files}{Vector of filter html files to crunch}

\item{species}{Vector of species names to crunch}

\item{ordered_regions}{Ordered vector of regions, as desired in output}

\item{filter_prefix}{Regular expression to remove from beginning of filter
names}

\item{filter_suffix}{Regular expression to remove from end of filter names}

\item{tax}{eBird taxonomy to use. Defaults to current taxonomy object from package
rebird}
}
\value{
Nested list containing structured filter information
}
\description{
Create nested data structure and crunch the filters
}
