% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebirdfiltercompare-package.R
\docType{package}
\name{ebirdfiltercompare-package}
\alias{ebirdfiltercompare}
\alias{ebirdfiltercompare-package}
\title{ebirdfiltercompare: Visualize Differences between eBird Filters}
\description{
This script is designed to be used by eBird Regional Editors to easily visualize differences among taxa filters in your regions.
}
\author{
\strong{Maintainer}: David Slager \email{slager@users.noreply.github.com} (\href{https://orcid.org/0000-0003-2525-2039}{ORCID})

Authors:
\itemize{
  \item Nick Anich
  \item Nicholas Walton
}

}
\keyword{internal}
