% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_pdf}
\alias{generate_pdf}
\title{Function to draw main PDF}
\usage{
generate_pdf(data, species, ordered_regions, output_directory = "output")
}
\arguments{
\item{data}{Nested list containing structured filter information}

\item{species}{Vector of species names present across all the filters}

\item{ordered_regions}{Ordered vector of regions, as desired in output}

\item{output_directory}{Directory to output PDF file. Defaults to 'output'.
Silently created if missing.}
}
\value{
Nothing. Side effect is taxa.pdf is written to output directory
}
\description{
Function to draw main PDF
}
