% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_filenames}
\alias{get_filenames}
\title{Get filter html filenames}
\usage{
get_filenames(
  path = system.file(file.path("extdata", "filter_htm"), package = "ebirdfiltercompare")
)
}
\arguments{
\item{path}{Path to check for filter html files. Defaults to package example.}
}
\value{
Vector of filter html file paths
}
\description{
Get filter html filenames
}
