% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_ordered_region_list}
\alias{import_ordered_region_list}
\title{Load custom-ordered regions created by user}
\usage{
import_ordered_region_list(
  file_path = system.file(file.path("extdata", "ordered_regions.csv"), package =
    "ebirdfiltercompare")
)
}
\arguments{
\item{file_path}{Path of manually ordered regions CSV file}
}
\value{
Vector of manually ordered region names
}
\description{
Load custom-ordered regions created by user
}
